+++ app/src/main/java/com/battlelancer/seriesguide/util/Metacritic.kt
@@ -1,5 +1,5 @@
-// Copyright 2023 Uwe Trottmann
 // SPDX-License-Identifier: Apache-2.0
+// Copyright 2020, 2023, 2024 Uwe Trottmann
 
 package com.battlelancer.seriesguide.util
 
@@ -12,7 +12,7 @@ object Metacritic {
      * Starts VIEW Intent with Metacritic website movie search results URL.
      */
     fun searchForMovie(context: Context, title: String) {
-        val url = "https://www.metacritic.com/search/movie/${Uri.encode(title)}/results"
+        val url = "https://www.metacritic.com/search/${Uri.encode(title)}/?category=2"
         WebTools.openInApp(context, url)
     }
 
@@ -20,7 +20,7 @@ object Metacritic {
      * Starts VIEW Intent with Metacritic website TV search results URL.
      */
     fun searchForTvShow(context: Context, title: String) {
-        val url = "https://www.metacritic.com/search/tv/${Uri.encode(title)}/results"
+        val url = "https://www.metacritic.com/search/${Uri.encode(title)}/?category=1"
         WebTools.openInApp(context, url)
     }