+++ app/src/main/java/com/battlelancer/seriesguide/util/tasks/AddListTask.java
@@ -1,5 +1,5 @@
-// Copyright 2023 Uwe Trottmann
 // SPDX-License-Identifier: Apache-2.0
+// Copyright 2016-2019, 2021, 2023 Uwe Trottmann
 
 package com.battlelancer.seriesguide.util.tasks;
 
@@ -8,6 +8,7 @@ import android.content.ContentValues;
 import android.content.Context;
 import androidx.annotation.NonNull;
 import androidx.annotation.VisibleForTesting;
+import com.battlelancer.seriesguide.R;
 import com.battlelancer.seriesguide.SgApp;
 import com.battlelancer.seriesguide.backend.HexagonTools;
 import com.battlelancer.seriesguide.provider.SeriesGuideContract;
@@ -95,6 +96,10 @@ public class AddListTask extends BaseActionTask {
 
     @Override
     protected int getSuccessTextResId() {
-        return 0; // display no success message
+        if (isSendingToHexagon()) {
+            return R.string.ack_list_added;
+        } else {
+            return 0;
+        }
     }
 }